move TotalDepthPixels to depthcamera, move ErrorTolerance to ZBuffer and rename to DepthCameraPrecisionUtil
